for compatibility with 2.7 and 3.x printing to stderr

+++ buildApp.py
@@ -29,6 +29,7 @@ Mac OSX
  
  to run from command line to catch errors: ./dist/tt.app/Contents/MacOS/tt
 '''
+from __future__ import print_function
 import sys, os, shutil
 
 if sys.platform == 'win32':
@@ -36,13 +37,14 @@ if sys.platform == 'win32':
 	os.system('python PyInstaller\pyinstaller.py --runtime-hook rthook_pyqt4.py -w -F -i timbretalk.ico --noupx -p . tt.py')
 elif sys.platform == 'darwin':
 	print ("Building app for Mac OSX")
-	shutil.rmtree('dist/tt.app') # save a step for builder by removing previous build
+	if os.path.isdir('dist/tt.app'):
+		shutil.rmtree('dist/tt.app') # save a step for builder by removing previous build
 	os.system('pyinstaller --clean --runtime-hook rthook_pyqt4.py -w -F -i timbretalk.icns  --noupx -p . --osx-bundle-identifier="TimbreTalk" tt.py')
 elif sys.platform[:5] == 'linux':
 	"Building app for Linux"
 	os.system('PyInstaller/pyinstaller.py --runtime-hook rthook_pyqt4.py -w -F -i timbretalk.ico -p . tt.py')
 else:
-	print ('unknown system platform: %s'%sys.platform, file=sys.stderr)
+	print('unknown system platform: %s'%sys.platfor, file=sys.stderr)
 	sys.exit()
 
 print ('Your application image is available in "dist/"')
